Reinitialize highlightjs in every markdown block
This is necessary because comments, etc., manipulate the DOM, and we
need to re-highlight when we do that because our new stuff might have
a code block in it.

This is all necessary because this plugins only interaction with the
page that is seen by the browser is the output of the transformation
from markdown to HTML -- so that output must contain the right
javascript to cause highlighting to happen.